@@ -7,7 +7,7 @@ description: A test fixture for webdev testing.
 # and build_web_compilers constraint should match those defined
 # in pubspec.dart.
 environment:
-  sdk: '>=2.11.99 <3.0.0'
+  sdk: '>=2.13.0-144.0.dev <3.0.0'
 
 dev_dependencies:
   build_runner: '>=1.6.2 <2.0.0'
