@@ -1,6 +1,6 @@
 name: webdev
 # Every time this changes you need to run `pub run build_runner build`.
-version: 2.7.2-dev
+version: 2.7.2
 # We should not depend on a dev SDK before publishing.
 # publish_to: none
 homepage: https://github.com/dart-lang/webdev
@@ -8,46 +8,44 @@ description: >-
   A CLI for Dart web development. Provides an easy and consistent set of
   features for users and tools to build and deploy web applications with Dart.
 
-# We require 2.11.99 as we require testing on 2.12 and above but are not
-# prepared to migrate this package to null-safety at this time.
 environment:
-  sdk: ">=2.11.99 <3.0.0"
+  sdk: ">=2.13.0-144.0.dev <3.0.0"
 
 dependencies:
-  args: ^1.5.0
+  args: ^2.0.0
   async: ^2.2.0
   build_daemon: ^2.0.0
   browser_launcher: ^0.1.5
-  crypto: ^2.0.6
+  crypto: ^3.0.0
   dds: ^1.4.1
-  dwds: ^7.0.0
-  http: ^0.12.0
-  http_multi_server: ^2.1.0
+  dwds: ^10.0.0
+  http: ^0.13.0
+  http_multi_server: ^3.0.0
   io: ^0.3.2+1
-  logging: ^0.11.0
+  logging: ^1.0.0
   pedantic: ^1.5.0
   meta: ^1.1.2
   path: ^1.5.1
   pool: ^1.4.0
-  pubspec_parse: ^0.1.4
-  pub_semver: ^1.3.2
-  shelf: ^0.7.4
-  shelf_proxy: ^0.1.0+5
-  shelf_static: ^0.2.8
+  pubspec_parse: ^1.0.0
+  pub_semver: ^2.0.0
+  shelf: ^1.0.0
+  shelf_proxy: ^1.0.0
+  shelf_static: ^1.0.0
   sse: ^3.7.0
   vm_service: '>=3.0.0 <7.0.0'
-  webkit_inspection_protocol: '>=0.4.0 <0.8.0'
-  yaml: ^2.1.13
+  webkit_inspection_protocol: '>=0.4.0 <2.0.0'
+  yaml: ^3.0.0
 
 dev_dependencies:
-  build: ^1.0.0
+  build: ^2.0.0
   build_runner: ^1.6.2
   build_verify: ^1.0.0
   build_version: ^2.0.0
   test: ^1.15.0
   test_descriptor: ^1.0.3
   test_process: ^1.0.1
-  webdriver: ^2.0.0
+  webdriver: ^3.0.0
 
 executables:
   webdev:
