@@ -2,11 +2,14 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// @dart = 2.9
+
 @JS()
 library background;
 
 import 'dart:async';
 import 'dart:convert';
+import 'dart:html';
 import 'dart:js';
 
 import 'package:async/async.dart';
@@ -14,6 +17,7 @@ import 'package:built_collection/built_collection.dart';
 import 'package:dwds/data/devtools_request.dart';
 import 'package:dwds/data/extension_request.dart';
 import 'package:dwds/data/serializers.dart';
+import 'package:dwds/src/sockets.dart';
 import 'package:js/js.dart';
 import 'package:js/js_util.dart' as js_util;
 import 'package:pedantic/pedantic.dart';
@@ -133,6 +137,9 @@ void main() {
         sendResponse(_tabIdToEncodedUri[request.tabId] ?? '');
       } else if (request.name == 'dwds.startDebugging') {
         startDebugging(null);
+        // TODO(grouma) - Actually determine if debugging initiated
+        // successfully.
+        sendResponse(true);
       } else {
         sendResponse(
             ErrorResponse()..error = 'Unknown request name: ${request.name}');
@@ -210,6 +217,30 @@ Future<void> _startSseClient(
   Tab currentTab,
   String dwdsVersion,
 ) async {
+  if (Version.parse(dwdsVersion ?? '0.0.0') >= Version.parse('9.1.0')) {
+    var authUri = uri.replace(path: authenticationPath);
+    if (authUri.scheme == 'ws') authUri = authUri.replace(scheme: 'http');
+    if (authUri.scheme == 'wss') authUri = authUri.replace(scheme: 'https');
+    var authUrl = authUri.toString();
+    try {
+      var response = await HttpRequest.request(authUrl,
+          method: 'GET', withCredentials: true);
+      if (!response.responseText
+          .contains('Dart Debug Authentication Success!')) {
+        throw Exception('Not authenticated.');
+      }
+    } catch (_) {
+      if (window.confirm(
+          'Authentication required.\n\nClick OK to authenticate then try again.')) {
+        // TODO(grouma) - see if we can get a callback on a successful auth
+        // and automatically reinitiate the dev workflow.
+        window.open(authUrl, 'Dart DevTools Authentication');
+        detach(Debuggee(tabId: currentTab.id), allowInterop(() {}));
+      }
+      return;
+    }
+  }
+
   // Specifies whether the debugger is attached.
   //
   // A debugger is detached if it is closed by user or the target is closed.
@@ -532,38 +563,3 @@ external set onFakeClick(void Function() f);
 
 @JS('window.isDartDebugExtension')
 external set isDartDebugExtension(_);
-
-abstract class SocketClient {
-  StreamSink<dynamic> get sink;
-  Stream<String> get stream;
-  void close();
-}
-
-class SseSocketClient extends SocketClient {
-  final SseClient _client;
-  SseSocketClient(this._client);
-
-  @override
-  StreamSink<dynamic> get sink => _client.sink;
-
-  @override
-  Stream<String> get stream => _client.stream;
-
-  @override
-  void close() => _client.close();
-}
-
-class WebSocketClient extends SocketClient {
-  final WebSocketChannel _channel;
-
-  WebSocketClient(this._channel);
-
-  @override
-  StreamSink<dynamic> get sink => _channel.sink;
-  @override
-  Stream<String> get stream =>
-      _channel.stream.map((dynamic o) => o?.toString());
-
-  @override
-  void close() => _channel.sink.close();
-}
