@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// @dart = 2.9
+
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
@@ -34,11 +36,17 @@ class DwdsInjector {
   final String _extensionUri;
   final _devHandlerPaths = StreamController<String>();
   final _logger = Logger('DwdsInjector');
+  final bool _enableDevtoolsLaunch;
+  final bool _useSseForInjectedClient;
 
   DwdsInjector(
     this._loadStrategy, {
     String extensionUri,
-  }) : _extensionUri = extensionUri;
+    bool enableDevtoolsLaunch,
+    bool useSseForInjectedClient,
+  })  : _extensionUri = extensionUri,
+        _enableDevtoolsLaunch = enableDevtoolsLaunch,
+        _useSseForInjectedClient = useSseForInjectedClient ?? true;
 
   /// Returns the embedded dev handler paths.
   ///
@@ -74,7 +82,12 @@ class DwdsInjector {
               var requestedUri = request.requestedUri;
               var appId = base64
                   .encode(md5.convert(utf8.encode('$requestedUri')).bytes);
-              var requestedUriBase = '${request.requestedUri.scheme}'
+              var scheme = '${request.requestedUri.scheme}';
+              if (!_useSseForInjectedClient) {
+                // Switch http->ws and https->wss.
+                scheme = scheme.replaceFirst('http', 'ws');
+              }
+              var requestedUriBase = '$scheme'
                   '://${request.requestedUri.authority}';
               var devHandlerPath = '\$dwdsSseHandler';
               var subPath = request.url.pathSegments.toList()..removeLast();
@@ -93,6 +106,7 @@ class DwdsInjector {
                 entrypoint,
                 _extensionUri,
                 _loadStrategy,
+                _enableDevtoolsLaunch,
               );
               body += await _loadStrategy.bootstrapFor(entrypoint);
               _logger.info('Injected debugging metadata for '
@@ -122,6 +136,7 @@ String _injectClientAndHoistMain(
   String entrypointPath,
   String extensionUri,
   LoadStrategy loadStrategy,
+  bool enableDevtoolsLaunch,
 ) {
   var bodyLines = body.split('\n');
   var extensionIndex =
@@ -140,6 +155,7 @@ String _injectClientAndHoistMain(
     entrypointPath,
     extensionUri,
     loadStrategy,
+    enableDevtoolsLaunch,
   );
   result += '''
   // Injected by dwds for debugging support.
@@ -172,6 +188,7 @@ String _injectedClientSnippet(
   String entrypointPath,
   String extensionUri,
   LoadStrategy loadStrategy,
+  bool enableDevtoolsLaunch,
 ) {
   var injectedBody = 'window.\$dartAppId = "$appId";\n'
       'window.\$dartReloadConfiguration = "${loadStrategy.reloadConfiguration}";\n'
@@ -179,6 +196,7 @@ String _injectedClientSnippet(
       'window.\$loadModuleConfig = ${loadStrategy.loadModuleSnippet};\n'
       'window.\$dwdsVersion = "$packageVersion";\n'
       'window.\$dwdsDevHandlerPath = "$devHandlerPath";\n'
+      'window.\$dwdsEnableDevtoolsLaunch = $enableDevtoolsLaunch;\n'
       'window.\$dartEntrypointPath = "$entrypointPath";\n'
       '${loadStrategy.loadClientSnippet(_clientScript)}';
   if (extensionUri != null) {
