@@ -2,14 +2,21 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// @dart = 2.9
+
 import 'dart:io';
 
 import 'package:async/async.dart';
 import 'package:http_multi_server/http_multi_server.dart';
+import 'package:shelf/shelf.dart';
 import 'package:shelf/shelf_io.dart';
 
+import '../../data/extension_request.dart';
 import '../handlers/socket_connections.dart';
 import 'extension_debugger.dart';
+
+const authenticationResponse = 'Dart Debug Authentication Success!\n\n'
+    'You can close this tab and launch the Dart Debug Extension again.';
 
 /// A backend for the Dart Debug Extension.
 ///
@@ -32,8 +39,18 @@ class ExtensionBackend {
   // Starts the backend on an open port.
   static Future<ExtensionBackend> start(
       SocketHandler _socketHandler, String hostname) async {
+    var cascade = Cascade();
+    cascade = cascade.add((request) {
+      if (request.url.path == authenticationPath) {
+        return Response.ok(authenticationResponse, headers: {
+          'Access-Control-Allow-Origin': request.headers['origin'],
+          'Access-Control-Allow-Credentials': 'true'
+        });
+      }
+      return Response.notFound('');
+    }).add(_socketHandler.handler);
     var server = await HttpMultiServer.bind(hostname, 0);
-    serveRequests(server, _socketHandler.handler);
+    serveRequests(server, cascade.handler);
     return ExtensionBackend._(
         _socketHandler, server.address.host, server.port, server);
   }
