@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// @dart = 2.9
+
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
@@ -168,13 +170,15 @@ class ChromeProxyService implements VmServiceInterface {
 
   Future<void> updateCompilerDependencies(String entrypoint) async {
     var metadataProvider = globalLoadStrategy.metadataProviderFor(entrypoint);
+    var moduleFormat = globalLoadStrategy.moduleFormat;
     var soundNullSafety = await metadataProvider.soundNullSafety;
 
     _logger.info('Initializing expression compiler for $entrypoint '
         'with sound null safety: $soundNullSafety');
 
     if (_compiler != null) {
-      await _compiler.initialize(soundNullSafety: soundNullSafety);
+      await _compiler?.initialize(
+          moduleFormat: moduleFormat, soundNullSafety: soundNullSafety);
       var dependencies =
           await globalLoadStrategy.moduleInfoForEntrypoint(entrypoint);
       var stopwatch = Stopwatch()..start();
@@ -196,26 +200,33 @@ class ChromeProxyService implements VmServiceInterface {
       throw UnsupportedError(
           'Cannot create multiple isolates for the same app');
     }
-
+    // Waiting for the debugger to be ready before initializing the entrypoint.
+    //
+    // Note: moving `await _debugger` after the `_initalizeEntryPoint` call
+    // causes `getcwd` system calls to fail. Since that system call is used
+    // in first `Uri.base` call in the expression compiler service isolate,
+    // the expression compiler service will fail to start.
+    // Issue: https://github.com/dart-lang/webdev/issues/1282
+    var debugger = await _debugger;
     await _initializeEntrypoint(appConnection.request.entrypointPath);
 
-    (await _debugger).notifyPausedAtStart();
+    debugger.notifyPausedAtStart();
     _inspector = await AppInspector.initialize(
       appConnection,
       remoteDebugger,
       _assetReader,
       _locations,
       uri,
-      await _debugger,
+      debugger,
       executionContext,
     );
 
-    await (await _debugger)
-        .reestablishBreakpoints(_previousBreakpoints, _disabledBreakpoints);
+    await debugger.reestablishBreakpoints(
+        _previousBreakpoints, _disabledBreakpoints);
     _disabledBreakpoints.clear();
 
     unawaited(appConnection.onStart.then((_) async {
-      await (await _debugger).resumeFromStart();
+      await debugger.resumeFromStart();
     }));
 
     var isolateRef = _inspector.isolateRef;
@@ -345,14 +356,77 @@ ${globalLoadStrategy.loadModuleSnippet}("dart_sdk").developer.invokeExtension(
     return _rpcNotSupportedFuture('clearVMTimeline');
   }
 
+  void _validateIsolateId(String isolateId) {
+    var isolate = _inspector?.isolate;
+    if (isolate?.id != isolateId) {
+      throw RPCError('evaluateInFrame', RPCError.kInvalidParams,
+          'Unrecognized isolate id: $isolateId. Supported isolate: ${isolate?.id}');
+    }
+  }
+
+  Future<Response> _getEvaluationResult(
+      Future<RemoteObject> Function() evaluation, String expression) async {
+    try {
+      var result = await evaluation();
+      // Handle compilation errors, internal errors,
+      // and reference errors from JavaScript evaluation in chrome.
+      if (result.type.contains('Error')) {
+        if (!result.type.startsWith('CompilationError')) {
+          _logger.warning('Failed to evaluate expression \'$expression\': '
+              '${result.type}: ${result.value}.');
+
+          _logger.info('Please follow instructions at '
+              'https://github.com/dart-lang/webdev/issues/956 '
+              'to file a bug.');
+        }
+        return ErrorRef(
+          kind: 'error',
+          message: '${result.type}: ${result.value}',
+          id: createId(),
+        );
+      }
+      return _inspector?.instanceHelper?.instanceRefFor(result);
+    } on RPCError catch (_) {
+      rethrow;
+    } catch (e, s) {
+      // Handle errors that throw exceptions, such as invalid JavaScript
+      // generated by the expression evaluator.
+      _logger.warning('Failed to evaluate expression \'$expression\'. ');
+      _logger.info('Please follow instructions at '
+          'https://github.com/dart-lang/webdev/issues/956 '
+          'to file a bug.');
+      _logger.info('$e:$s');
+      return ErrorRef(kind: 'error', message: '<unknown>', id: createId());
+    }
+  }
+
   @override
   Future<Response> evaluate(
-      String isolateId, String targetId, String expression,
-      {Map<String, String> scope, bool disableBreakpoints}) async {
+    String isolateId,
+    String targetId,
+    String expression, {
+    Map<String, String> scope,
+    bool disableBreakpoints,
+  }) async {
     // TODO(798) - respect disableBreakpoints.
     var stopwatch = Stopwatch()..start();
     dynamic error;
+
     try {
+      if (_expressionEvaluator != null) {
+        _validateIsolateId(isolateId);
+
+        var library = await _inspector?.getLibrary(isolateId, targetId);
+        var result = await _getEvaluationResult(
+            () => _expressionEvaluator.evaluateExpression(
+                isolateId, library.uri, expression),
+            expression);
+        if (result is ErrorRef) {
+          error = result;
+        }
+        return result;
+      }
+      // fall back to javascript evaluation
       var remote = await _inspector?.evaluate(isolateId, targetId, expression,
           scope: scope);
       return _inspector?.instanceHelper?.instanceRefFor(remote);
@@ -373,58 +447,38 @@ ${globalLoadStrategy.loadModuleSnippet}("dart_sdk").developer.invokeExtension(
   Future<Response> evaluateInFrame(
       String isolateId, int frameIndex, String expression,
       {Map<String, String> scope, bool disableBreakpoints}) async {
-    if (_expressionEvaluator != null) {
-      var isolate = _inspector?.isolate;
-      if (isolate?.id != isolateId) {
-        throw RPCError('evaluateInFrame', RPCError.kInvalidParams,
-            'Unrecognized isolate id: $isolateId. Supported isolate: ${isolate?.id}');
-      }
-      dynamic error;
-      var stopwatch = Stopwatch()..start();
-      try {
-        var result = await _expressionEvaluator.evaluateExpression(
-            isolateId, frameIndex, expression);
+    // TODO(798) - respect disableBreakpoints.
 
-        // Handle compilation errors, internal errors,
-        // and reference errors from JavaScript evaluation in chrome
-        if (result.type.contains('Error')) {
-          if (!result.type.startsWith('CompilationError')) {
-            _logger.warning('Failed to evaluate expression \'$expression\': '
-                '${result.type}: ${result.value}.');
+    var stopwatch = Stopwatch()..start();
+    dynamic error;
 
-            _logger.info('Please follow instructions at '
-                'https://github.com/dart-lang/webdev/issues/956 '
-                'to file a bug.');
-          }
-          return ErrorRef(
-            kind: 'error',
-            message: '${result.type}: ${result.value}',
-            id: createId(),
-          );
+    try {
+      if (_expressionEvaluator != null) {
+        _validateIsolateId(isolateId);
+
+        var result = await _getEvaluationResult(
+            () => _expressionEvaluator.evaluateExpressionInFrame(
+                isolateId, frameIndex, expression),
+            expression);
+
+        if (result is ErrorRef) {
+          error = result;
         }
-        return _inspector?.instanceHelper?.instanceRefFor(result);
-      } catch (e, s) {
-        error = e;
-        // Handle errors that throw exceptions, such as invalid JavaScript
-        // generated by the expression evaluator
-        _logger.warning('Failed to evaluate expression \'$expression\'. ');
-        _logger.info('Please follow instructions at '
-            'https://github.com/dart-lang/webdev/issues/956 '
-            'to file a bug.');
-        _logger.info('$e:$s');
-        return ErrorRef(kind: 'error', message: '<unknown>', id: createId());
-      } finally {
-        emitEvent(DwdsEvent('EVALUATE_IN_FRAME', {
-          'expression': expression,
-          'success': error == null,
-          'exception': error,
-          'elapsedMilliseconds': stopwatch.elapsedMilliseconds,
-        }));
+        return result;
       }
+      throw RPCError('evaluateInFrame', RPCError.kInvalidRequest,
+          'Expression evaluation is not supported for this configuration.');
+    } catch (e) {
+      error = e;
+      rethrow;
+    } finally {
+      emitEvent(DwdsEvent('EVALUATE_IN_FRAME', {
+        'expression': expression,
+        'success': error == null,
+        'exception': error,
+        'elapsedMilliseconds': stopwatch.elapsedMilliseconds,
+      }));
     }
-
-    throw RPCError('evaluateInFrame', RPCError.kInvalidRequest,
-        'Expression evaluation is not supported for this configuration.');
   }
 
   @override
@@ -457,8 +511,8 @@ ${globalLoadStrategy.loadModuleSnippet}("dart_sdk").developer.invokeExtension(
     var isolate = _inspector?.isolate;
     if (isolate?.id == isolateId) return isolate;
     // TODO: Throw an RPC error here.
-    throw ArgumentError.value(
-        isolateId, 'isolateId', 'Unrecognized isolate id');
+    throw ArgumentError.value(isolateId, 'isolateId',
+        'Unrecognized isolate id. (Supported isolate: ${isolate?.id})');
   }
 
   @override
