@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// @dart = 2.9
+
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
@@ -60,7 +62,7 @@ void main() async {
       service =
           ExpressionCompilerService('localhost', port, assetHandler, false);
 
-      await service.initialize();
+      await service.initialize(moduleFormat: 'amd');
 
       // setup asset server
       server = await HttpMultiServer.bind('localhost', port);
