@@ -2,7 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-@Tags(['frontend-server'])
+// @dart = 2.9
+
 @TestOn('vm')
 import 'dart:async';
 
