@@ -2,7 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-@Tags(['frontend-server'])
+// @dart = 2.9
+
 @TestOn('vm')
 import 'dart:async';
 
@@ -16,17 +17,43 @@ import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 import 'fixtures/context.dart';
 import 'fixtures/logging.dart';
 
-final context = TestContext(
-    directory: p.join('..', 'fixtures', '_testPackage'),
-    entry: p.join('..', 'fixtures', '_testPackage', 'web', 'main.dart'),
-    path: 'index.html',
-    pathToServe: 'web');
+class TestSetup {
+  static final contextUnsound = TestContext(
+      directory: p.join('..', 'fixtures', '_testPackage'),
+      entry: p.join('..', 'fixtures', '_testPackage', 'web', 'main.dart'),
+      path: 'index.html',
+      pathToServe: 'web');
 
-ChromeProxyService get service =>
-    fetchChromeProxyService(context.debugConnection);
-WipConnection get tabConnection => context.tabConnection;
+  TestContext context;
+
+  TestSetup.unsound() : context = contextUnsound;
+
+  ChromeProxyService get service =>
+      fetchChromeProxyService(context.debugConnection);
+  WipConnection get tabConnection => context.tabConnection;
+}
 
 void main() async {
+  var setup = TestSetup.unsound();
+  var context = setup.context;
+
+  Future<void> onBreakPoint(String isolate, ScriptRef script,
+      String breakPointId, Future<void> Function() body) async {
+    Breakpoint bp;
+    try {
+      var line =
+          await context.findBreakpointLine(breakPointId, isolate, script);
+      bp = await setup.service
+          .addBreakpointWithScriptUri(isolate, script.uri, line);
+      await body();
+    } finally {
+      // Remove breakpoint so it doesn't impact other tests or retries.
+      if (bp != null) {
+        await setup.service.removeBreakpoint(isolate, bp.id);
+      }
+    }
+  }
+
   group('shared context with evaluation', () {
     setUpAll(() async {
       // Note: change 'printOnFailure' to 'print' for debug printing
@@ -53,168 +80,178 @@ void main() async {
       Stream<Event> stream;
 
       setUp(() async {
-        vm = await service.getVM();
-        isolate = await service.getIsolate(vm.isolates.first.id);
-        scripts = await service.getScripts(isolate.id);
+        vm = await setup.service.getVM();
+        isolate = await setup.service.getIsolate(vm.isolates.first.id);
+        scripts = await setup.service.getScripts(isolate.id);
 
-        await service.streamListen('Debug');
-        stream = service.onEvent('Debug');
+        await setup.service.streamListen('Debug');
+        stream = setup.service.onEvent('Debug');
 
         mainScript = scripts.scripts
             .firstWhere((each) => each.uri.contains('main.dart'));
       });
 
       tearDown(() async {
-        await service.resume(isolate.id);
+        await setup.service.resume(isolate.id);
       });
 
       test('local', () async {
-        var line = await context.findBreakpointLine(
-            'printLocal', isolate.id, mainScript);
-        var bp = await service.addBreakpointWithScriptUri(
-            isolate.id, mainScript.uri, line);
+        await onBreakPoint(isolate.id, mainScript, 'printLocal', () async {
+          var event = await stream.firstWhere(
+              (Event event) => event.kind == EventKind.kPauseBreakpoint);
 
-        var event = await stream.firstWhere(
-            (Event event) => event.kind == EventKind.kPauseBreakpoint);
+          var result = await setup.service
+              .evaluateInFrame(isolate.id, event.topFrame.index, 'local');
 
-        var result = await service.evaluateInFrame(
-            isolate.id, event.topFrame.index, 'local');
-
-        expect(
-            result,
-            const TypeMatcher<InstanceRef>().having(
-                (instance) => instance.valueAsString, 'valueAsString', '42'));
-
-        // Remove breakpoint so it doesn't impact other tests.
-        await service.removeBreakpoint(isolate.id, bp.id);
+          expect(
+              result,
+              const TypeMatcher<InstanceRef>().having(
+                  (instance) => instance.valueAsString, 'valueAsString', '42'));
+        });
       });
 
       test('global', () async {
-        var line = await context.findBreakpointLine(
-            'printGlobal', isolate.id, mainScript);
-        var bp = await service.addBreakpointWithScriptUri(
-            isolate.id, mainScript.uri, line);
+        await onBreakPoint(isolate.id, mainScript, 'printGlobal', () async {
+          var event = await stream.firstWhere(
+              (Event event) => event.kind == EventKind.kPauseBreakpoint);
 
-        var event = await stream.firstWhere(
-            (Event event) => event.kind == EventKind.kPauseBreakpoint);
+          var result = await setup.service.evaluateInFrame(
+              isolate.id, event.topFrame.index, 'testLibraryValue');
 
-        var result = await service.evaluateInFrame(
-            isolate.id, event.topFrame.index, 'testLibraryValue');
-
-        expect(
-            result,
-            const TypeMatcher<InstanceRef>().having(
-                (instance) => instance.valueAsString, 'valueAsString', '3'));
-
-        // Remove breakpoint so it doesn't impact other tests.
-        await service.removeBreakpoint(isolate.id, bp.id);
+          expect(
+              result,
+              const TypeMatcher<InstanceRef>().having(
+                  (instance) => instance.valueAsString, 'valueAsString', '3'));
+        });
       });
 
       test('call core function', () async {
-        var line = await context.findBreakpointLine(
-            'printLocal', isolate.id, mainScript);
-        var bp = await service.addBreakpointWithScriptUri(
-            isolate.id, mainScript.uri, line);
+        await onBreakPoint(isolate.id, mainScript, 'printLocal', () async {
+          var event = await stream.firstWhere(
+              (Event event) => event.kind == EventKind.kPauseBreakpoint);
 
-        var event = await stream.firstWhere(
-            (Event event) => event.kind == EventKind.kPauseBreakpoint);
+          var result = await setup.service.evaluateInFrame(
+              isolate.id, event.topFrame.index, 'print(local)');
 
-        var result = await service.evaluateInFrame(
-            isolate.id, event.topFrame.index, 'print(local)');
-
-        expect(
-            result,
-            const TypeMatcher<InstanceRef>().having(
-                (instance) => instance.valueAsString, 'valueAsString', 'null'));
-
-        // Remove breakpoint so it doesn't impact other tests.
-        await service.removeBreakpoint(isolate.id, bp.id);
+          expect(
+              result,
+              const TypeMatcher<InstanceRef>().having(
+                  (instance) => instance.valueAsString,
+                  'valueAsString',
+                  'null'));
+        });
       });
 
       test('call library function with const param', () async {
-        var line = await context.findBreakpointLine(
-            'printLocal', isolate.id, mainScript);
-        var bp = await service.addBreakpointWithScriptUri(
-            isolate.id, mainScript.uri, line);
+        await onBreakPoint(isolate.id, mainScript, 'printLocal', () async {
+          var event = await stream.firstWhere(
+              (Event event) => event.kind == EventKind.kPauseBreakpoint);
 
-        var event = await stream.firstWhere(
-            (Event event) => event.kind == EventKind.kPauseBreakpoint);
+          var result = await setup.service.evaluateInFrame(
+              isolate.id, event.topFrame.index, 'testLibraryFunction(42)');
 
-        var result = await service.evaluateInFrame(
-            isolate.id, event.topFrame.index, 'testLibraryFunction(42)');
-
-        expect(
-            result,
-            const TypeMatcher<InstanceRef>().having(
-                (instance) => instance.valueAsString, 'valueAsString', '42'));
-
-        // Remove breakpoint so it doesn't impact other tests.
-        await service.removeBreakpoint(isolate.id, bp.id);
+          expect(
+              result,
+              const TypeMatcher<InstanceRef>().having(
+                  (instance) => instance.valueAsString, 'valueAsString', '42'));
+        });
       });
 
       test('call library function with local param', () async {
-        var line = await context.findBreakpointLine(
-            'printLocal', isolate.id, mainScript);
-        var bp = await service.addBreakpointWithScriptUri(
-            isolate.id, mainScript.uri, line);
+        await onBreakPoint(isolate.id, mainScript, 'printLocal', () async {
+          var event = await stream.firstWhere(
+              (Event event) => event.kind == EventKind.kPauseBreakpoint);
 
-        var event = await stream.firstWhere(
-            (Event event) => event.kind == EventKind.kPauseBreakpoint);
+          var result = await setup.service.evaluateInFrame(
+              isolate.id, event.topFrame.index, 'testLibraryFunction(local)');
 
-        var result = await service.evaluateInFrame(
-            isolate.id, event.topFrame.index, 'testLibraryFunction(local)');
+          expect(
+              result,
+              const TypeMatcher<InstanceRef>().having(
+                  (instance) => instance.valueAsString, 'valueAsString', '42'));
+        });
+      });
+
+      test('loop variable', () async {
+        await onBreakPoint(isolate.id, mainScript, 'printLoopVariable',
+            () async {
+          var event = await stream.firstWhere(
+              (Event event) => event.kind == EventKind.kPauseBreakpoint);
+
+          var result = await setup.service
+              .evaluateInFrame(isolate.id, event.topFrame.index, 'item');
+
+          expect(
+              result,
+              const TypeMatcher<InstanceRef>().having(
+                  (instance) => instance.valueAsString, 'valueAsString', '1'));
+        });
+      });
+
+      test('error', () async {
+        await onBreakPoint(isolate.id, mainScript, 'printLocal', () async {
+          var event = await stream.firstWhere(
+              (Event event) => event.kind == EventKind.kPauseBreakpoint);
+
+          var error = await setup.service
+              .evaluateInFrame(isolate.id, event.topFrame.index, 'typo');
+
+          expect(
+              error,
+              const TypeMatcher<ErrorRef>().having(
+                  (instance) => instance.message,
+                  'message',
+                  'CompilationError: Getter not found: \'typo\'.\ntypo\n^^^^'));
+        });
+      });
+    });
+
+    group('evaluate', () {
+      VM vm;
+      Isolate isolate;
+
+      setUp(() async {
+        vm = await setup.service.getVM();
+        isolate = await setup.service.getIsolate(vm.isolates.first.id);
+
+        await setup.service.streamListen('Debug');
+      });
+
+      tearDown(() async {});
+
+      test('uses symbol from the same library', () async {
+        var library = isolate.rootLib;
+        var result = await setup.service
+            .evaluate(isolate.id, library.id, 'MainClass(0).toString()');
+
+        expect(
+            result,
+            const TypeMatcher<InstanceRef>().having(
+                (instance) => instance.valueAsString, 'valueAsString', '0'));
+      });
+
+      test('uses symbol from another library', () async {
+        var library = isolate.rootLib;
+        var result = await setup.service.evaluate(
+            isolate.id, library.id, 'TestLibraryClass(0,1).toString()');
+
+        expect(
+            result,
+            const TypeMatcher<InstanceRef>().having(
+                (instance) => instance.valueAsString,
+                'valueAsString',
+                'field: 0, _field: 1'));
+      });
+
+      test('closure call', () async {
+        var library = isolate.rootLib;
+        var result = await setup.service
+            .evaluate(isolate.id, library.id, '(() => 42)()');
 
         expect(
             result,
             const TypeMatcher<InstanceRef>().having(
                 (instance) => instance.valueAsString, 'valueAsString', '42'));
-
-        // Remove breakpoint so it doesn't impact other tests.
-        await service.removeBreakpoint(isolate.id, bp.id);
-      });
-
-      test('loop variable', () async {
-        var line = await context.findBreakpointLine(
-            'printLoopVariable', isolate.id, mainScript);
-        var bp = await service.addBreakpointWithScriptUri(
-            isolate.id, mainScript.uri, line);
-
-        var event = await stream.firstWhere(
-            (Event event) => event.kind == EventKind.kPauseBreakpoint);
-
-        var result = await service.evaluateInFrame(
-            isolate.id, event.topFrame.index, 'item');
-
-        expect(
-            result,
-            const TypeMatcher<InstanceRef>().having(
-                (instance) => instance.valueAsString, 'valueAsString', '1'));
-
-        // Remove breakpoint so it doesn't impact other tests.
-        await service.removeBreakpoint(isolate.id, bp.id);
-      }, tags: ['dev-sdk']);
-
-      test('error', () async {
-        var line = await context.findBreakpointLine(
-            'printLocal', isolate.id, mainScript);
-        var bp = await service.addBreakpointWithScriptUri(
-            isolate.id, mainScript.uri, line);
-
-        var event = await stream.firstWhere(
-            (Event event) => event.kind == EventKind.kPauseBreakpoint);
-
-        var error = await service.evaluateInFrame(
-            isolate.id, event.topFrame.index, 'typo');
-
-        expect(
-            error,
-            const TypeMatcher<ErrorRef>().having(
-                (instance) => instance.message,
-                'message',
-                'CompilationError: Getter not found: \'typo\'.\ntypo\n^^^^'));
-
-        // Remove breakpoint so it doesn't impact other tests.
-        await service.removeBreakpoint(isolate.id, bp.id);
       });
     });
   });
@@ -239,37 +276,31 @@ void main() async {
       Stream<Event> stream;
 
       setUp(() async {
-        vm = await service.getVM();
-        isolate = await service.getIsolate(vm.isolates.first.id);
-        scripts = await service.getScripts(isolate.id);
+        vm = await setup.service.getVM();
+        isolate = await setup.service.getIsolate(vm.isolates.first.id);
+        scripts = await setup.service.getScripts(isolate.id);
 
-        await service.streamListen('Debug');
-        stream = service.onEvent('Debug');
+        await setup.service.streamListen('Debug');
+        stream = setup.service.onEvent('Debug');
 
         mainScript = scripts.scripts
             .firstWhere((each) => each.uri.contains('main.dart'));
       });
 
       tearDown(() async {
-        await service.resume(isolate.id);
+        await setup.service.resume(isolate.id);
       });
 
       test('cannot evaluate expression', () async {
-        var line = await context.findBreakpointLine(
-            'printLocal', isolate.id, mainScript);
-        var bp = await service.addBreakpointWithScriptUri(
-            isolate.id, mainScript.uri, line);
+        await onBreakPoint(isolate.id, mainScript, 'printLocal', () async {
+          var event = await stream.firstWhere(
+              (Event event) => event.kind == EventKind.kPauseBreakpoint);
 
-        var event = await stream.firstWhere(
-            (Event event) => event.kind == EventKind.kPauseBreakpoint);
-
-        expect(
-            () => service.evaluateInFrame(
-                isolate.id, event.topFrame.index, 'local'),
-            throwsRPCError);
-
-        // Remove breakpoint so it doesn't impact other tests.
-        await service.removeBreakpoint(isolate.id, bp.id);
+          expect(
+              () => setup.service
+                  .evaluateInFrame(isolate.id, event.topFrame.index, 'local'),
+              throwsRPCError);
+        });
       });
     });
   });
