@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// @dart = 2.9
+
 @JS()
 library require_reloading_manager;
 
@@ -144,7 +146,7 @@ class RequireRestarter implements Restarter {
   }
 
   List<String> _moduleParents(String module) =>
-      requireLoader.moduleParentsGraph.get(module)?.cast();
+      requireLoader.moduleParentsGraph.get(module)?.cast() ?? [];
 
   int _moduleTopologicalCompare(String module1, String module2) {
     var topological = 0;
