@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+// @dart = 2.9
+
 // Note: this is a copy from flutter tools, updated to work with dwds tests
 
 import 'dart:async';
@@ -11,11 +13,11 @@ import 'dart:io';
 import 'package:dwds/dwds.dart';
 import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
+import 'package:package_config/package_config.dart';
 import 'package:path/path.dart' as p;
 import 'package:pedantic/pedantic.dart';
 import 'package:usage/uuid/uuid.dart';
 
-import 'package_map.dart';
 import 'utilities.dart';
 
 String get frontendServerExecutable =>
@@ -157,18 +159,24 @@ class StdoutHandler {
 class PackageUriMapper {
   PackageUriMapper(String scriptPath, String packagesFilePath,
       String fileSystemScheme, List<String> fileSystemRoots) {
-    var packageMap = PackageMap(fileSystem.path.absolute(packagesFilePath)).map;
+    init(scriptPath, packagesFilePath, fileSystemScheme, fileSystemRoots);
+  }
+
+  Future<void> init(String scriptPath, String packagesFilePath,
+      String fileSystemScheme, List<String> fileSystemRoots) async {
+    var packageConfig = await loadPackageConfig(
+        File(fileSystem.path.absolute(packagesFilePath)));
     var isWindowsPath =
         Platform.isWindows && !scriptPath.startsWith('org-dartlang-app');
     var scriptUri = Uri.file(scriptPath, windows: isWindowsPath).toString();
-    for (var packageName in packageMap.keys) {
-      var prefix = packageMap[packageName].toString();
+    for (var package in packageConfig.packages) {
+      var prefix = package.packageUriRoot.toString();
       // Only perform a multi-root mapping if there are multiple roots.
       if (fileSystemScheme != null &&
           fileSystemRoots != null &&
           fileSystemRoots.length > 1 &&
           prefix.contains(fileSystemScheme)) {
-        _packageName = packageName;
+        _packageName = package.name;
         _uriPrefixes = fileSystemRoots
             .map((String name) =>
                 Uri.file(name, windows: Platform.isWindows).toString())
@@ -176,7 +184,7 @@ class PackageUriMapper {
         return;
       }
       if (scriptUri.startsWith(prefix)) {
-        _packageName = packageName;
+        _packageName = package.name;
         _uriPrefixes = <String>[prefix];
         return;
       }
@@ -779,5 +787,5 @@ class TestExpressionCompiler implements ExpressionCompiler {
       true;
 
   @override
-  Future<void> initialize({bool soundNullSafety}) async {}
+  Future<void> initialize({String moduleFormat, bool soundNullSafety}) async {}
 }
