@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// @dart = 2.9
+
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
@@ -197,7 +199,8 @@ class _PackageInfo {
 
 /// Returns the package info for the latest webdev release.
 Future<_PackageInfo> _latestPackageInfo() async {
-  var response = await get('https://pub.dartlang.org/api/packages/webdev',
+  var response = await get(
+      Uri.parse('https://pub.dartlang.org/api/packages/webdev'),
       headers: {HttpHeaders.userAgentHeader: 'webdev $packageVersion'});
   var responseObj = json.decode(response.body);
   var pubspec = Pubspec.fromJson(
