@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// @dart = 2.9
+
 @Timeout(Duration(minutes: 5))
 import 'dart:io';
 
@@ -276,117 +278,226 @@ void main() {
     for (var soundNullSafety in [false, true]) {
       var nullSafetyOption = soundNullSafety ? 'sound' : 'unsound';
       group('--null-safety=$nullSafetyOption', () {
-        test('and --enable-expression-evaluation', () async {
-          var openPort = await findUnusedPort();
-          // running daemon command that starts dwds without keyboard input
-          var args = [
-            'daemon',
-            'web:$openPort',
-            '--enable-expression-evaluation',
-            '--verbose',
-          ];
-          var process = await runWebDev(args,
-              workingDirectory:
-                  soundNullSafety ? soundExampleDirectory : exampleDirectory);
-          VmService vmService;
+        group('and --enable-expression-evaluation:', () {
+          test('evaluateInFrame', () async {
+            var openPort = await findUnusedPort();
+            // running daemon command that starts dwds without keyboard input
+            var args = [
+              'daemon',
+              'web:$openPort',
+              '--enable-expression-evaluation',
+              '--verbose',
+            ];
+            var process = await runWebDev(args,
+                workingDirectory:
+                    soundNullSafety ? soundExampleDirectory : exampleDirectory);
+            VmService vmService;
 
-          try {
-            // Wait for debug service Uri
-            String wsUri;
-            await expectLater(process.stdout, emitsThrough((message) {
-              wsUri = getDebugServiceUri(message as String);
-              return wsUri != null;
-            }));
-            expect(wsUri, isNotNull);
+            try {
+              // Wait for debug service Uri
+              String wsUri;
+              await expectLater(process.stdout, emitsThrough((message) {
+                wsUri = getDebugServiceUri(message as String);
+                return wsUri != null;
+              }));
+              expect(wsUri, isNotNull);
 
-            vmService = await vmServiceConnectUri(wsUri);
-            var vm = await vmService.getVM();
-            var isolate = vm.isolates.first;
-            var scripts = await vmService.getScripts(isolate.id);
+              vmService = await vmServiceConnectUri(wsUri);
+              var vm = await vmService.getVM();
+              var isolate = vm.isolates.first;
+              var scripts = await vmService.getScripts(isolate.id);
 
-            await vmService.streamListen('Debug');
-            var stream = vmService.onEvent('Debug');
+              await vmService.streamListen('Debug');
+              var stream = vmService.onEvent('Debug');
 
-            var mainScript = scripts.scripts
-                .firstWhere((each) => each.uri.contains('main.dart'));
+              var mainScript = scripts.scripts
+                  .firstWhere((each) => each.uri.contains('main.dart'));
 
-            var bpLine = await findBreakpointLine(
-                vmService, 'printCounter', isolate.id, mainScript);
+              var bpLine = await findBreakpointLine(
+                  vmService, 'printCounter', isolate.id, mainScript);
 
-            var bp = await vmService.addBreakpointWithScriptUri(
-                isolate.id, mainScript.uri, bpLine);
-            expect(bp, isNotNull);
+              var bp = await vmService.addBreakpointWithScriptUri(
+                  isolate.id, mainScript.uri, bpLine);
+              expect(bp, isNotNull);
 
-            await stream.firstWhere(
-                (Event event) => event.kind == EventKind.kPauseBreakpoint);
+              await stream.firstWhere(
+                  (Event event) => event.kind == EventKind.kPauseBreakpoint);
 
-            var result = await vmService.evaluateInFrame(isolate.id, 0, 'true');
-            expect(
-                result,
-                const TypeMatcher<InstanceRef>().having(
-                    (instance) => instance.valueAsString,
-                    'valueAsString',
-                    'true'));
-          } finally {
-            await vmService?.dispose();
-            await exitWebdev(process);
-            await process.shouldExit();
-          }
-        }, timeout: const Timeout.factor(2));
+              var result =
+                  await vmService.evaluateInFrame(isolate.id, 0, 'true');
+              expect(
+                  result,
+                  const TypeMatcher<InstanceRef>().having(
+                      (instance) => instance.valueAsString,
+                      'valueAsString',
+                      'true'));
+            } finally {
+              await vmService?.dispose();
+              await exitWebdev(process);
+              await process.shouldExit();
+            }
+          }, timeout: const Timeout.factor(2));
 
-        test('and --no-enable-expression-evaluation', () async {
-          var openPort = await findUnusedPort();
-          var args = [
-            'daemon',
-            'web:$openPort',
-            '--no-enable-expression-evaluation',
-          ];
-          var process = await runWebDev(args,
-              workingDirectory:
-                  soundNullSafety ? soundExampleDirectory : exampleDirectory);
-          VmService vmService;
+          test('evaluate', () async {
+            var openPort = await findUnusedPort();
+            // running daemon command that starts dwds without keyboard input
+            var args = [
+              'daemon',
+              'web:$openPort',
+              '--enable-expression-evaluation',
+              '--verbose',
+            ];
+            var process = await runWebDev(args,
+                workingDirectory:
+                    soundNullSafety ? soundExampleDirectory : exampleDirectory);
+            VmService vmService;
 
-          try {
-            // Wait for debug service Uri
-            String wsUri;
-            await expectLater(process.stdout, emitsThrough((message) {
-              wsUri = getDebugServiceUri(message as String);
-              return wsUri != null;
-            }));
-            expect(wsUri, isNotNull);
+            try {
+              // Wait for debug service Uri
+              String wsUri;
+              await expectLater(process.stdout, emitsThrough((message) {
+                wsUri = getDebugServiceUri(message as String);
+                return wsUri != null;
+              }));
+              expect(wsUri, isNotNull);
 
-            vmService = await vmServiceConnectUri(wsUri);
-            var vm = await vmService.getVM();
-            var isolate = vm.isolates.first;
-            var scripts = await vmService.getScripts(isolate.id);
+              vmService = await vmServiceConnectUri(wsUri);
+              var vm = await vmService.getVM();
+              var isolate = await vmService.getIsolate(vm.isolates.first.id);
+              var library = isolate.rootLib;
 
-            await vmService.streamListen('Debug');
-            var stream = vmService.onEvent('Debug');
+              await vmService.streamListen('Debug');
 
-            var mainScript = scripts.scripts
-                .firstWhere((each) => each.uri.contains('main.dart'));
+              var result = await vmService.evaluate(isolate.id, library.id,
+                  '(document?.body?.children?.first as SpanElement)?.text');
 
-            var bpLine = await findBreakpointLine(
-                vmService, 'printCounter', isolate.id, mainScript);
+              expect(
+                  result,
+                  const TypeMatcher<InstanceRef>().having(
+                      (instance) => instance.valueAsString,
+                      'valueAsString',
+                      'Hello World!!'));
 
-            var bp = await vmService.addBreakpointWithScriptUri(
-                isolate.id, mainScript.uri, bpLine);
-            expect(bp, isNotNull);
+              result = await vmService.evaluate(
+                  isolate.id, library.id, 'main.toString()');
 
-            var event = await stream.firstWhere(
-                (Event event) => event.kind == EventKind.kPauseBreakpoint);
+              expect(
+                  result,
+                  const TypeMatcher<InstanceRef>().having(
+                      (instance) => instance.valueAsString,
+                      'valueAsString',
+                      contains('Hello World!!')));
+            } finally {
+              await vmService?.dispose();
+              await exitWebdev(process);
+              await process.shouldExit();
+            }
+          }, timeout: const Timeout.factor(2));
+        });
 
-            expect(
-                () => vmService.evaluateInFrame(
-                    isolate.id, event.topFrame.index, 'true'),
-                throwsRPCError);
-          } finally {
-            await vmService?.dispose();
-            await exitWebdev(process);
-            await process.shouldExit();
-          }
+        group('and --no-enable-expression-evaluation:', () {
+          test('evaluateInFrame', () async {
+            var openPort = await findUnusedPort();
+            var args = [
+              'daemon',
+              'web:$openPort',
+              '--no-enable-expression-evaluation',
+              '--verbose',
+            ];
+            var process = await runWebDev(args,
+                workingDirectory:
+                    soundNullSafety ? soundExampleDirectory : exampleDirectory);
+            VmService vmService;
+
+            try {
+              // Wait for debug service Uri
+              String wsUri;
+              await expectLater(process.stdout, emitsThrough((message) {
+                wsUri = getDebugServiceUri(message as String);
+                return wsUri != null;
+              }));
+              expect(wsUri, isNotNull);
+
+              vmService = await vmServiceConnectUri(wsUri);
+              var vm = await vmService.getVM();
+              var isolate = vm.isolates.first;
+              var scripts = await vmService.getScripts(isolate.id);
+
+              await vmService.streamListen('Debug');
+              var stream = vmService.onEvent('Debug');
+
+              var mainScript = scripts.scripts
+                  .firstWhere((each) => each.uri.contains('main.dart'));
+
+              var bpLine = await findBreakpointLine(
+                  vmService, 'printCounter', isolate.id, mainScript);
+
+              var bp = await vmService.addBreakpointWithScriptUri(
+                  isolate.id, mainScript.uri, bpLine);
+              expect(bp, isNotNull);
+
+              var event = await stream.firstWhere(
+                  (Event event) => event.kind == EventKind.kPauseBreakpoint);
+
+              expect(
+                  () => vmService.evaluateInFrame(
+                      isolate.id, event.topFrame.index, 'true'),
+                  throwsRPCError);
+            } finally {
+              await vmService?.dispose();
+              await exitWebdev(process);
+              await process.shouldExit();
+            }
+          });
+
+          test('evaluate', () async {
+            var openPort = await findUnusedPort();
+            // running daemon command that starts dwds without keyboard input
+            var args = [
+              'daemon',
+              'web:$openPort',
+              '--no-enable-expression-evaluation',
+              '--verbose',
+            ];
+            var process = await runWebDev(args,
+                workingDirectory:
+                    soundNullSafety ? soundExampleDirectory : exampleDirectory);
+            VmService vmService;
+
+            try {
+              // Wait for debug service Uri
+              String wsUri;
+              await expectLater(process.stdout, emitsThrough((message) {
+                wsUri = getDebugServiceUri(message as String);
+                return wsUri != null;
+              }));
+              expect(wsUri, isNotNull);
+
+              vmService = await vmServiceConnectUri(wsUri);
+              var vm = await vmService.getVM();
+              var isolate = await vmService.getIsolate(vm.isolates.first.id);
+              var library = isolate.rootLib;
+
+              await vmService.streamListen('Debug');
+
+              // Evaluate falls back to js evaluation
+              var result = await vmService.evaluate(
+                  isolate.id, library.id, 'main.toString()');
+
+              expect(
+                  result,
+                  const TypeMatcher<InstanceRef>().having(
+                      (instance) => instance.valueAsString,
+                      'valueAsString',
+                      contains('Hello World!!')));
+            } finally {
+              await vmService?.dispose();
+              await exitWebdev(process);
+              await process.shouldExit();
+            }
+          }, timeout: const Timeout.factor(2));
         });
       });
     }
-  }, tags: 'expression-compilation-service');
+  });
 }
